requirements.txt
pymodbus
pyserial
